fix: 4dp 구분선 위치 수정 (PlaceWaitingSection 외부 섹션 구분자로 이동)
